requirements.txt
google-cloud-vision
argparse
piexif
Pillow
